+++ crypto/tls/tls.go
@@ -5,10 +5,12 @@ import (
 	"crypto/x509"
 	"flag"
 	"os"
+	"strings"
 
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
+	cliFlags "k8s.io/component-base/cli/flag"
 )
 
 // ClientConfig is the config for client TLS.
@@ -18,6 +20,8 @@ type ClientConfig struct {
 	CAPath             string `yaml:"tls_ca_path" category:"advanced"`
 	ServerName         string `yaml:"tls_server_name" category:"advanced"`
 	InsecureSkipVerify bool   `yaml:"tls_insecure_skip_verify" category:"advanced"`
+	CipherSuites       string `yaml:"tls_cipher_suites" category:"advanced"`
+	MinVersion         string `yaml:"tls_min_version" category:"advanced"`
 }
 
 var (
@@ -32,6 +36,8 @@ func (cfg *ClientConfig) RegisterFlagsWithPrefix(prefix string, f *flag.FlagSet)
 	f.StringVar(&cfg.CAPath, prefix+".tls-ca-path", "", "Path to the CA certificates file to validate server certificate against. If not set, the host's root CA certificates are used.")
 	f.StringVar(&cfg.ServerName, prefix+".tls-server-name", "", "Override the expected name on the server certificate.")
 	f.BoolVar(&cfg.InsecureSkipVerify, prefix+".tls-insecure-skip-verify", false, "Skip validating server certificate.")
+	f.StringVar(&cfg.CipherSuites, prefix+".tls-cipher-suites", "", "Override the default cipher suite list (separated by commas).")
+	f.StringVar(&cfg.MinVersion, prefix+".tls-min-version", "", "Override the default minimum TLS version.")
 }
 
 // GetTLSConfig initialises tls.Config from config options
@@ -67,6 +73,23 @@ func (cfg *ClientConfig) GetTLSConfig() (*tls.Config, error) {
 			return nil, errors.Wrapf(err, "failed to load TLS certificate %s,%s", cfg.CertPath, cfg.KeyPath)
 		}
 		config.Certificates = []tls.Certificate{clientCert}
+	}
+
+	if cfg.MinVersion != "" {
+		minVersion, err := cliFlags.TLSVersion(cfg.MinVersion)
+		if err != nil {
+			return nil, errors.Wrapf(err, "failed to set minimum TLS version %s", cfg.MinVersion)
+		}
+		config.MinVersion = minVersion
+	}
+
+	if cfg.CipherSuites != "" {
+		rawCipherSuites := strings.Split(cfg.CipherSuites, ",")
+		cipherSuites, err := cliFlags.TLSCipherSuites(rawCipherSuites)
+		if err != nil {
+			return nil, errors.Wrapf(err, "failed to set cipher suites %s", cfg.CipherSuites)
+		}
+		config.CipherSuites = cipherSuites
 	}
 
 	return config, nil