+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+* [ENHANCEMENT] Added `-.tls-min-version` and `.tls-cipher-suites` flag. #217
 * [CHANGE] Added new `-consul.cas-retry-delay` flag. It has a default value of `1s`, while previously there was no delay between retries. #178
 * [CHANGE] Flagext: `DayValue` now always uses UTC when parsing or displaying dates. #71
 * [CHANGE] Closer: remove the closer package since it's trivial to just copy/paste. #70